+++ src/IEnte.java
@@ -1,4 +1,4 @@
 public interface IEnte {
     public boolean isAlive();
-    public int getHP();
+    public float getHP();
 }
